Add more tests for bn128 proof gen

+++ src/mnt6753.js
@@ -796,7 +796,7 @@ class Mnt6753 {
             this.putBin(pr, rnd);
         } else {
             br = NodeCrypto.randomBytes(16);
-            this.putBin(pr, new Uint32Array(br.buffer));
+            this.putBin(pr, new Uint8Array(br.buffer));
         }
